#23: move data logic from form- to globalSlice
