+++ frontend/src/store/reducers.tsx
@@ -5,8 +5,10 @@
 import { combineReducers } from '@reduxjs/toolkit'
 import formControllerSlice from '../components/form/formControllerSlice'
 import counterReducer from '../components/sampleCounter/counterSlice'
+import globalSlice from './globalSlice'
 
 export default combineReducers({
+	globalControl: globalSlice,
 	formControl: formControllerSlice,
 	counter: counterReducer,
 })
